Fix problem with detecting if device is closest to home with multiple devices
The problem was that if you have to compare your changed device with two
other devices you overwrote the device_is_closest_to_home when you
compared it with each device.
for example:
comparing it to the first device might set device_is_closest_to_home to
false and comparing it with the second device might set it back to true.

now you assume the changed device is closest to home and only change it
to false when you find a device that is closer to home. this will also
remove the need to check for the amount of devices you compared it with,
when you can't compare it with other devices, it's automatically the
closest device already :)